Add files via upload
As simple C# Unity script to capture screenshots from the game view window. It works in both play mode and outside of it. If you provide a folder name (e.g., "Screenshots"), it will save images in that folder. If the folder doesn’t exist, it will be created automatically. If left empty, images will be stored in the default path (the Assets folder).
PS: Make sure to refresh " ctrl+r " after taking a screenshot